// src/app/form/types/interface.ts
export interface KPIElements {
    geographyId: number;
    locationName: string;
    year: number;
    kpiValue: number;
    natureOfData: string;
    existingValue: number;
    population: number;
    kpiName: string;
    kpiLabel: string;
    subCategory1: string;
    subCategory2: string;
    subCategory3: string;
}